feat: adds docker setup for git proxy
Introduces Dockerfiles, compose files, and a setup script
to streamline the deployment and configuration of Git Proxy
in containerized environments.

The setup script automates the generation of SSH keys, TLS
certificates (optional), and a Docker-specific config file,
simplifying the initial setup process. It also allows
selection between development and production compose
configurations.

The Dockerfiles use a multi-stage build to optimize the
final image size, and the compose files provide configurations
for both development and production environments.